requirements.txt
feedparser
yt-dlp
google-generativeai
openai-whisper
opencc-python-reimplemented
